requirements.txt
discord.py[voice]
discord-py-interactions
python-dotenv
